@@ -2,38 +2,36 @@
 name = "orapa-mine"
 version = "1.0.0"
 description = "Orapa Mine board game - multiplayer web implementation"
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 dependencies = [
-    "fastapi>=0.109.0",
-    "uvicorn[standard]>=0.27.0",
-    "python-multipart>=0.0.6",
-    "piccolo[postgres]>=1.11.0",
-    "psycopg[binary]>=3.1.0",
-    "asyncpg>=0.29.0",
-    "python-jose[cryptography]>=3.3.0",
+    "asyncpg>=0.30.0",
+    "fastapi>=0.120.1",
     "passlib[bcrypt]>=1.7.4",
-    "python-dotenv>=1.0.0",
-    "websockets>=12.0",
-    "pydantic>=2.5.3",
-    "pydantic-settings>=2.1.0",
+    "piccolo[postgres]>=1.30.0",
+    "psycopg[binary]>=3.2.12",
+    "pydantic>=2.12.3",
+    "pydantic-settings>=2.11.0",
+    "python-dotenv>=1.2.1",
+    "python-jose[cryptography]>=3.5.0",
+    "python-multipart>=0.0.20",
+    "uvicorn[standard]>=0.38.0",
+    "websockets>=15.0.1",
 ]
 
-[project.optional-dependencies]
+
+[dependency-groups]
 dev = [
-    "pytest==7.4.3",
-    "pytest-asyncio==0.21.1",
-    "httpx==0.26.0",
-    "ruff==0.1.14",
-    "mypy==1.8.0",
+    "httpx>=0.28.1",
+    "mypy>=1.18.2",
+    "poethepoet>=0.37.0",
+    "pytest>=8.4.2",
+    "pytest-asyncio>=1.2.0",
+    "ruff>=0.14.2",
 ]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
 
 [tool.ruff]
 line-length = 100
-target-version = "py311"
+target-version = "py312"
 
 [tool.ruff.lint]
 select = [
@@ -48,13 +46,13 @@ select = [
     "SIM", # flake8-simplify
 ]
 ignore = [
-    "E501",  # line too long (handled by formatter)
-    "B008",  # do not perform function calls in argument defaults
-    "W191",  # indentation contains tabs
+    "E501", # line too long (handled by formatter)
+    "B008", # do not perform function calls in argument defaults
+    "W191", # indentation contains tabs
 ]
 
 [tool.ruff.lint.per-file-ignores]
-"__init__.py" = ["F401"]  # Unused imports in __init__.py
+"__init__.py" = ["F401"] # Unused imports in __init__.py
 
 [tool.ruff.format]
 quote-style = "double"
@@ -63,7 +61,7 @@ skip-magic-trailing-comma = false
 line-ending = "auto"
 
 [tool.mypy]
-python_version = "3.11"
+python_version = "3.12"
 strict = true
 warn_return_any = true
 warn_unused_configs = true
@@ -81,11 +79,7 @@ follow_imports = "normal"
 ignore_missing_imports = true
 
 [[tool.mypy.overrides]]
-module = [
-    "piccolo.*",
-    "passlib.*",
-    "jose.*",
-]
+module = ["piccolo.*", "passlib.*", "jose.*"]
 ignore_missing_imports = true
 
 [tool.pytest.ini_options]
@@ -94,3 +88,33 @@ python_files = ["test_*.py"]
 python_classes = ["Test*"]
 python_functions = ["test_*"]
 asyncio_mode = "auto"
+
+[tool.poe.tasks]
+# Linting
+lint = "ruff check ./app"
+lint-fix = "ruff check --fix ./app"
+
+# Formatting
+format = "ruff format ./app"
+format-check = "ruff format --check ./app"
+
+# Type checking
+type = "mypy ./app"
+
+env = "cp .env.example .env"
+
+run = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
+
+migrate = "piccolo migrations forwards app"
+
+clean = "find . -type f -name '*.pyc' -delete && find . -type d -name '__pycache__' -delete"
+
+# Combined checks
+[tool.poe.tasks.check]
+help = "Run all checks (lint, format, type)"
+sequence = ["format-check", "lint", "type"]
+ignore_fail = false
+
+[tool.poe.tasks.fix]
+help = "Auto-fix linting and formatting issues"
+sequence = ["format", "lint-fix"]
