@@ -1,10 +1,12 @@
 """
 Orapa Mine FastAPI Application
 """
+
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from app.config import settings
+
 from app.api import api_router
+from app.config import settings
 
 app = FastAPI(
     title=settings.app_name,
@@ -29,11 +31,7 @@ app.include_router(api_router, prefix="/api/v1")
 @app.get("/")
 async def root() -> dict:
     """Root endpoint."""
-    return {
-        "name": settings.app_name,
-        "version": settings.app_version,
-        "status": "running"
-    }
+    return {"name": settings.app_name, "version": settings.app_version, "status": "running"}
 
 
 @app.get("/health")
@@ -44,9 +42,5 @@ async def health_check() -> dict:
 
 if __name__ == "__main__":
     import uvicorn
-    uvicorn.run(
-        "app.main:app",
-        host=settings.host,
-        port=settings.port,
-        reload=settings.debug
-    )
+
+    uvicorn.run("app.main:app", host=settings.host, port=settings.port, reload=settings.debug)
