@@ -1,37 +1,34 @@
 """
 Game-related models for Orapa Mine.
 """
-from piccolo.table import Table
-from piccolo.columns import (
-    Varchar,
-    Integer,
-    Boolean,
-    Text,
-    Timestamp,
-    JSON,
-    ForeignKey,
-)
-from piccolo.columns.readable import Readable
+
 from enum import Enum
+
+from piccolo.columns import JSON, Boolean, ForeignKey, Integer, Text, Timestamp, Varchar
+from piccolo.columns.readable import Readable
+from piccolo.table import Table
 
 
 class GameStatus(str, Enum):
     """Game status enum."""
-    WAITING = "waiting"       # Waiting for players
-    SETUP = "setup"           # Director is placing pieces
+
+    WAITING = "waiting"  # Waiting for players
+    SETUP = "setup"  # Director is placing pieces
     IN_PROGRESS = "in_progress"  # Game is active
-    COMPLETED = "completed"   # Game finished
-    CANCELLED = "cancelled"   # Game was cancelled
+    COMPLETED = "completed"  # Game finished
+    CANCELLED = "cancelled"  # Game was cancelled
 
 
 class PlayerRole(str, Enum):
     """Player role in a game."""
+
     DIRECTOR = "director"
     EXPLORER = "explorer"
 
 
 class PieceType(str, Enum):
     """Types of tangram pieces."""
+
     LARGE_TRIANGLE = "large_triangle"
     MEDIUM_TRIANGLE = "medium_triangle"
     SMALL_TRIANGLE = "small_triangle"
@@ -42,6 +39,7 @@ class PieceType(str, Enum):
 
 class PieceColor(str, Enum):
     """Colors of pieces."""
+
     RED = "red"
     BLUE = "blue"
     YELLOW = "yellow"
@@ -54,6 +52,7 @@ class Game(Table, tablename="games"):
     """
     Main game session table.
     """
+
     session_id = Varchar(length=50, unique=True, index=True)
     status = Varchar(length=20, default=GameStatus.WAITING.value)  # GameStatus
     current_turn_player = Varchar(length=50, null=True)  # Username of current player
@@ -74,6 +73,7 @@ class GamePlayer(Table, tablename="game_players"):
     """
     Players in a specific game session.
     """
+
     game = ForeignKey(Game, on_delete="CASCADE", index=True)
     username = Varchar(length=50, index=True)
     role = Varchar(length=20)  # PlayerRole
@@ -90,6 +90,7 @@ class Piece(Table, tablename="pieces"):
     """
     Available piece types (reference table).
     """
+
     piece_type = Varchar(length=50, unique=True)  # PieceType
     piece_color = Varchar(length=50)  # PieceColor
     area_cells = Integer()  # Number of cells the piece occupies
@@ -105,6 +106,7 @@ class PlacedPiece(Table, tablename="placed_pieces"):
     """
     Pieces placed by the director in a game.
     """
+
     game = ForeignKey(Game, on_delete="CASCADE", index=True)
     piece_type = Varchar(length=50)  # PieceType
     piece_color = Varchar(length=50)  # PieceColor
@@ -117,8 +119,7 @@ class PlacedPiece(Table, tablename="placed_pieces"):
     @classmethod
     def get_readable(cls):
         return Readable(
-            template="%s at (%s, %s)",
-            columns=[cls.piece_type, cls.position_x, cls.position_y]
+            template="%s at (%s, %s)", columns=[cls.piece_type, cls.position_x, cls.position_y]
         )
 
 
@@ -126,6 +127,7 @@ class WaveShot(Table, tablename="wave_shots"):
     """
     Record of elastic wave shots by explorers.
     """
+
     game = ForeignKey(Game, on_delete="CASCADE", index=True)
     player_username = Varchar(length=50, index=True)
     entry_position = Varchar(length=5)  # e.g., "5", "11", "E", "K"
@@ -138,5 +140,5 @@ class WaveShot(Table, tablename="wave_shots"):
     def get_readable(cls):
         return Readable(
             template="%s -> %s (%s)",
-            columns=[cls.entry_position, cls.exit_position, cls.exit_color]
+            columns=[cls.entry_position, cls.exit_position, cls.exit_color],
         )
