@@ -1,8 +1,9 @@
 """
 Game services and logic for Orapa Mine.
 """
-from app.services.piece_geometry import PieceGeometry, get_piece_geometry
+
 from app.services.color_mixer import ColorMixer
+from app.services.piece_geometry import PieceGeometry, get_piece_geometry
 from app.services.ray_tracer import RayTracer
 
 __all__ = [
